requirements.txt
beautifulsoup4==4.8.0
bokeh==1.3.4
boto==2.49.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==6.7
cloudinary==1.20.0
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.7
cssselect==1.1.0
Cython==0.29.13
cytoolz==0.10.0
datetime==4.3
dask==2.5.2
decorator==4.4.0
dropbox==10.1.2
docutils==0.15.2
email==6.0.0a1
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.0.2
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
get==2019.4.13
gevent==1.4.0
glob2==0.7
greenlet==0.4.15
gunicorn==20.0.4
h5py==2.9.0
HeapDict==1.0.1
html5lib==1.0.1
httplib2==0.11.3
hyperlink==19.0.0
idna==2.6
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==0.24
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
lazy==1.3
libarchive-c==2.8
lxml==4.4.1
MarkupSafe==1.0
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
networkx==2.3
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.16.5
numpydoc==0.9.1
oauth2==1.9.0.post1
oauthlib==2.0.7
olefile==0.46
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.16
psutil==5.6.3
psycopg2==2.8.5
public==2019.4.13
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycurl==7.43.0.3
PyDispatcher==2.0.5
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==2.0.2
PyJWT==1.4.2
pylint==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytz==2019.3
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
query-string==2019.4.13
queuelib==1.5.0
redis==3.5.1
request==2019.4.13
requests==2.23.0
requests-oauthlib==0.8.0
rope==0.14.0
rq==1.4.0
ruamel-yaml==0.15.46
schedule==0.6.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
Sijax==0.3.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
statsmodels==0.10.1
stripe==2.43.0
sympy==1.4
tables==3.5.2
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
Twisted==20.3.0
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.22
URLObject==2.4.3
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
WTForms==2.1
xlrd==1.2.0
xlwt==1.3.0
zict==1.0.0
zipp==0.6.0
zope.interface==5.1.0
